refactor: extract board config module
